requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
boto3==1.34.104
botocore==1.34.104
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-storages==1.14.3
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jmespath==1.0.1
Markdown==3.6
mysqlclient==2.2.4
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
